Added Zip Tests
