+++ tools/coinbase-call-faucet/.tool-dump.test.json
@@ -1,0 +1,1 @@
+{"type":"Deno","content":[{"toolkit_name":"coinbase-call-faucet","name":"Coinbase Faucet Caller","author":"Shinkai","version":"1.0.0","js_code":"import { Coinbase, CoinbaseOptions } from 'npm:@coinbase/coinbase-sdk@0.0.16';\n\ntype Configurations = {\n  name: string;\n  privateKey: string;\n  walletId?: string;\n};\ntype Parameters = {};\ntype Result = {\n  data: string;\n};\nexport type Run<C extends Record<string, any>, I extends Record<string, any>, R extends Record<string, any>> = (config: C, inputs: I) => Promise<R>;\n\nexport const run: Run<Configurations, Parameters, Result> = async (\n  configurations,\n  _params,\n): Promise<Result> => {\n  const coinbaseOptions: CoinbaseOptions = {\n    apiKeyName: configurations.name,\n    privateKey: configurations.privateKey,\n  };\n  const coinbase = new Coinbase(coinbaseOptions);\n  console.log(`Coinbase configured: `, coinbase);\n  const user = await coinbase.getDefaultUser();\n  console.log(`User: `, user);\n\n  // Use walletId from Config only\n  const walletId = configurations.walletId;\n\n  let wallet;\n  if (walletId) {\n    // Retrieve existing Wallet using walletId\n    wallet = await user.getWallet(walletId);\n    console.log(`Wallet retrieved: `, wallet.toString());\n  } else {\n    // Create a new Wallet for the User\n    wallet = await user.createWallet({\n      networkId: Coinbase.networks.BaseSepolia,\n    });\n    console.log(`Wallet successfully created: `, wallet.toString());\n  }\n\n  const faucetTransaction = await wallet.faucet();\n  console.log(\n    `Faucet transaction completed successfully: `,\n    faucetTransaction.toString(),\n  );\n\n  return {\n    data: `Faucet transaction completed successfully: ${faucetTransaction.toString()} for wallet: ${wallet.getDefaultAddress()}`,\n  };\n};\n\n","tools":[],"config":[{"BasicConfig":{"key_name":"name","description":"","required":true,"type":null,"key_value":null}},{"BasicConfig":{"key_name":"privateKey","description":"","required":true,"type":null,"key_value":null}},{"BasicConfig":{"key_name":"walletId","description":"","required":false,"type":null,"key_value":null}}],"description":"Tool for calling a faucet on Coinbase","keywords":["coinbase","faucet","shinkai"],"input_args":{"type":"object","properties":{},"required":[]},"output_arg":{"json":""},"activated":false,"embedding":[-0.99122494,0.50806594,-0.18137962,-0.41493967,-0.23193666,-0.46692222,0.0687634,0.43874425,-0.27778542,-0.017959813,-0.8661224,0.07842385,0.17415565,0.059652843,0.3853139,-0.53799176,0.011051909,-0.6348554,-1.2048376,0.28303137,0.50286824,0.9636021,0.029742166,0.2970993,-0.093243614,0.1579739,-0.29110375,0.12320799,-1.1625351,-2.4252186,0.5984823,1.1908039,-0.6606316,-0.2709986,-0.63542795,-0.34321687,0.12860778,-0.26192757,-0.15083013,-0.029253855,-0.03655629,0.2297895,-0.5603324,-0.050662898,-0.13666894,-0.70836246,0.48023468,0.19298565,0.53484815,0.7338745,0.5589467,-0.36540076,0.25469062,-0.5136771,-0.3332828,-0.12868685,0.22334099,-0.7157119,0.19501099,0.31472236,0.22343026,0.40181735,-3.8011842,-0.047244344,1.2716331,-0.6118786,-0.11626663,-0.11116794,-0.33865005,0.4694708,-0.037501752,-0.19524391,-0.4614096,0.3223101,0.29280674,0.037641026,0.7293865,-0.35483155,0.567163,0.12906854,0.13266593,0.6007808,0.12596978,0.43191862,-0.25101435,0.4072931,0.21287161,-0.52972436,0.29522818,0.6254314,-0.4135732,0.12478796,-0.078617595,0.07535772,0.000089384615,0.5190583,0.10056962,0.24503843,0.017940115,3.1678724,0.6107782,0.14412847,0.31150937,-0.9736704,0.5321445,-0.13874264,-0.30686077,-0.35969603,0.23723474,0.10950215,0.79324293,-0.023657102,0.6210422,0.19991389,0.34732723,-0.1934203,-0.13354202,0.36694792,-0.6958847,0.9096159,-0.48663473,0.664064,-1.2650294,-0.29368094,0.4261583,-0.076165274,-0.32999966,0.3633916,0.00080637634,0.13501783,0.9418749,-0.82700145,-0.76627177,-0.07468449,-0.28829494,0.434925,0.07580288,-0.6897222,-0.68812186,-0.40245715,0.18199688,-2.1058512,1.224085,-0.19669126,0.44360998,-0.17754836,0.4740172,0.1371968,-0.7740964,0.2550662,-0.17048314,0.36711055,0.42428753,-0.23090535,0.28756404,0.14982949,-0.35205346,0.062007338,-0.01289238,0.44881707,-0.42916048,0.2266951,0.28846586,0.23304105,0.5841187,-0.67397416,0.6724802,0.16376093,-0.3693137,-0.123955846,0.6306007,-0.07195953,-0.2700342,0.22918184,-1.4100586,-0.1133481,-0.27751988,0.05546531,0.30617833,-0.78547466,-0.46278316,0.99444,-0.53480095,-0.73826486,-0.08605997,0.055162385,0.06365322,0.22701722,0.21728194,0.5684299,-0.9437571,1.0672885,-0.35899705,-0.04806394,-0.29208297,0.21088497,-0.17880356,0.1196495,-0.049828578,-0.036527902,-0.27911696,-0.52558833,-0.34787142,0.22506356,-0.25697565,-0.13310519,1.3069135,0.49249384,0.12590644,-0.8827312,0.44646263,-0.06894559,0.42643484,0.7206083,0.8160076,-0.44844735,0.16804968,0.019308368,0.16603923,0.26426148,1.0348673,0.5334502,-0.954048,-0.39132386,-0.81579703,0.08473987,0.17618689,0.4596308,-0.5402172,0.21049465,0.31503275,0.7458434,0.6645061,0.84952253,0.8682018,0.7328625,-0.0679008,0.33489862,0.22054482,-0.12859799,-0.005194355,0.2583303,-0.6760137,0.5455183,0.2866763,-0.0079629645,0.42053515,-0.9807938,0.3063334,1.6347826,0.21526441,-0.5904196,0.11962591,-0.062463757,-0.050052047,-0.20544973,-1.1212342,0.5382988,-0.6786197,0.1453031,-0.40478265,-0.1791695,0.37327555,0.026863594,-0.25588816,-0.20552187,-0.37325567,0.1906169,0.117457904,-0.062925085,0.10161585,0.8090012,0.629976,-0.18731183,-0.12044861,0.19006206,0.31053948,0.17830685,-0.3020076,-0.5735512,-0.020475678,-0.16555937,0.6612981,-0.10536747,0.182735,-0.6303363,-0.5657179,0.18669339,-0.8254714,-0.26674804,-0.10148545,-0.34947407,-0.20487763,0.21583214,2.2093656,0.7205735,-0.16089511,0.3484195,0.35745317,0.91434324,0.17402184,-0.0005370714,-0.525771,0.16789791,-1.0236385,0.0937128,0.34643108,-0.49626967,-0.1790568,-0.32317,-0.37175724,0.30956614,-0.014280874,0.106248505,-0.20281261,-0.42128444,0.21055475,0.45975873,-0.2581095,-2.0881133,-0.26736298,0.31235436,-0.6298407,-0.12596682,0.24670933,0.6193566,0.19237164,0.60334843,-0.22912377,1.361868,0.35932636,-0.4044572,-0.9850079,0.4134363,0.5290532,0.45526838,0.1147489,-0.3668976,-0.42578667,0.3019122,-0.13172133,1.3124754,-0.020972705,0.30414128,-0.41859457,-0.37522227,-0.014403075,-1.7559369,0.020438597,0.16517067,-0.49357662,0.31749943,-0.03856969,-0.6470938,0.5298628,0.80309176,-0.2739704,-0.12535317,-0.0002693422,0.57529545,0.32525396,0.051337495,-0.58472085,-0.19945246,0.20148355,-0.051555976,0.31528756,0.27569154,0.088488206,0.26962948,-0.63993686,-0.3265281,0.13676783,-0.2961794,0.8000029,0.56923926,-0.22789086,-0.08803365,0.67435783,0.011274636,-0.16963835,0.25483674,-0.7386935,0.30054727],"result":{"type":"object","properties":{"data":{"type":"string"}},"required":["data"]},"sql_tables":null,"sql_queries":null,"file_inbox":null,"oauth":null,"assets":null},false]}
