+++ tools/coinbase-create-wallet/.tool-dump.test.json
@@ -1,0 +1,1 @@
+{"type":"Deno","content":[{"toolkit_name":"coinbase-create-wallet","name":"Coinbase Wallet Creator","author":"Shinkai","version":"1.0.0","js_code":"import { Coinbase, CoinbaseOptions } from 'npm:@coinbase/coinbase-sdk@0.0.16';\n\ntype Configurations = {\n  name: string;\n  privateKey: string;\n  useServerSigner?: string;\n};\ntype Parameters = {}; // Params type is now empty\ntype Result = {\n  walletId?: string;\n  seed?: string;\n  address?: string;\n};\nexport type Run<C extends Record<string, any>, I extends Record<string, any>, R extends Record<string, any>> = (config: C, inputs: I) => Promise<R>;\n\nexport const run: Run<Configurations, Parameters, Result> = async (\n  configurations: Configurations,\n  _parameters: Parameters,\n): Promise<Result> => {\n  const coinbaseOptions: CoinbaseOptions = {\n    apiKeyName: configurations.name,\n    privateKey: configurations.privateKey,\n    useServerSigner: configurations.useServerSigner === 'true',\n  };\n  const coinbase = new Coinbase(coinbaseOptions);\n  console.log(`Coinbase configured: `, coinbase);\n  const user = await coinbase.getDefaultUser();\n  console.log(`User: `, user);\n\n  // Create a new Wallet for the User\n  const wallet = await user.createWallet({\n    networkId: Coinbase.networks.BaseSepolia,\n  });\n  console.log(`Wallet successfully created: `, wallet.toString());\n\n  let exportedWallet;\n  if (!configurations.useServerSigner) {\n    exportedWallet = await wallet.export();\n  }\n\n  const address = await wallet.getDefaultAddress();\n\n  return {\n    ...exportedWallet,\n    walletId: wallet.getId(),\n    address: address?.getId(),\n  };\n};\n","tools":[],"config":[{"BasicConfig":{"key_name":"name","description":"","required":true,"type":null,"key_value":null}},{"BasicConfig":{"key_name":"privateKey","description":"","required":true,"type":null,"key_value":null}},{"BasicConfig":{"key_name":"useServerSigner","description":"","required":false,"type":null,"key_value":null}}],"description":"Tool for creating a Coinbase wallet","keywords":["coinbase","wallet","creator","shinkai"],"input_args":{"type":"object","properties":{},"required":[]},"output_arg":{"json":""},"activated":false,"embedding":[0.106608376,0.41127837,-0.25463364,-0.32248706,-0.104817785,0.121706314,-0.03151182,0.31224874,-0.2831944,-0.0034338525,-0.19127713,-0.008223534,0.22638485,-0.09930447,-0.28486118,-0.63315773,-0.09636158,-0.61439145,-1.1654487,0.38228172,-0.17458896,0.8497766,-0.43269163,0.15335917,-0.28330773,-0.2640039,0.0034893155,-0.5224032,-1.484486,-2.4849114,0.8048678,1.0833544,-0.5426029,-0.5558562,-0.05434633,-0.31097323,-0.17190091,0.5118817,-0.42390314,-0.31034046,0.35211498,0.27447006,-0.65473056,0.023174979,-0.058990005,-0.44352376,0.54819924,-0.39887503,0.26282844,0.40640676,-0.0002581142,-0.47739455,0.6547748,-0.03268323,-0.35610425,0.12967776,-0.26997164,-0.3305793,0.10562223,0.4433401,-0.040558998,0.330962,-3.8495562,0.22852053,1.3840722,0.042035088,0.50890285,-0.054913018,-0.3909843,0.62089247,-0.010075815,-0.11408972,-0.54617864,0.043472372,0.38645452,-0.1113597,0.75728303,-0.10065475,0.62765956,0.011410739,0.06718382,0.77804315,0.2650589,-0.18410273,-0.42870727,0.22244582,-0.1088205,-0.53857994,0.04582644,0.0972573,0.08662674,0.45195132,0.2139351,-0.111990795,-0.47918832,-0.25548273,0.016256634,0.9064493,0.47019023,2.697148,0.24152085,0.2536752,0.34584814,-1.0840755,0.79801327,-0.23469439,0.08639613,-0.4534926,0.16087207,-0.2777297,0.9485559,-0.2849575,0.1207515,0.256539,0.35246587,0.32869238,-0.04498246,0.14251262,-0.4105745,0.4183277,-0.26015264,0.2240419,-1.085044,-0.045694605,-0.14072339,-0.07030669,-0.4911205,0.2826326,-0.3474514,-0.5368858,0.21524921,-0.41232717,-0.43492627,0.008838819,-0.040434144,0.6099907,0.01878146,-0.66687995,-0.43180692,-0.02428811,0.30693483,-0.7558724,0.5635787,-0.18748997,0.08487825,-0.37317407,0.3038798,-0.035463236,-0.51118016,-0.600597,-0.010352377,-0.0957617,0.6385061,-0.30879894,0.64587414,-0.53867763,0.020915512,-0.5058636,-0.46977317,-0.086883,-0.22218241,-0.23072702,0.25326255,1.0340672,0.7886596,-0.31496307,0.48080894,0.27734283,0.5118715,-0.70576495,0.2801963,-0.117037974,0.4662382,0.15260935,-0.5982343,0.050490092,-0.41207424,-0.026747614,0.35468322,-0.36062858,-0.7574638,0.34588605,-0.93115747,-1.0822469,-0.7088281,0.3006152,0.4463941,0.22230771,0.10920119,0.72157216,-0.46959072,1.7862909,-0.026602164,-0.046276525,-0.101544656,0.14876622,-0.2731639,0.37518954,0.24139847,-0.28620204,-0.2476773,-0.30634525,-0.19084437,0.22665034,-0.20038769,-0.4942695,0.4065581,0.8478193,-0.1943736,-0.07214823,0.09502381,0.1875737,0.83186984,0.91354287,1.1263794,-0.18895322,0.5766026,-0.037720766,1.3292009,0.14854592,0.63262135,0.24933234,-0.77803963,-0.33052447,-0.7650778,-0.37940335,0.5418016,0.051896956,-0.3578351,0.0016623922,0.6015654,0.628299,0.61673456,0.4733826,1.2288805,0.5030629,-0.010545418,0.03804444,0.17318018,-1.260278,-0.14013255,0.27815554,-0.54101306,-0.7199203,-0.16819221,-0.7132447,0.21671882,-0.6276087,0.088249184,2.020304,0.36894408,-0.14315441,0.38600516,-0.21624562,-0.038657956,-0.18285182,-1.4090341,0.3883453,-0.4969522,0.4531218,-0.09483522,0.1403723,-0.12007758,0.21613769,-0.33294532,-0.4450294,-0.9099992,-0.23876816,-0.11038546,-0.38953996,0.057502277,1.1098955,0.6361157,0.4614854,-0.09154616,-0.015582925,0.70769066,0.46275842,-0.40131098,-0.046300434,0.111293405,-0.08412773,0.76974195,0.3433804,0.35255334,-0.5403275,-0.28918272,0.06919776,-0.79234564,-0.15432379,-0.087078966,-0.38518324,-0.5070818,0.6790647,1.6665665,1.2250792,1.1262922,0.41495696,-0.09596672,0.31750906,-0.13302247,0.5684142,0.096760765,0.17608225,-1.3890605,-0.37464488,0.40351182,-0.61909026,0.044795968,0.110169545,0.26001665,0.50215644,-0.0068575554,0.39997515,0.019612342,-0.45005175,-0.20951176,0.5002105,-0.27554086,-2.350778,-0.23506549,-0.13005003,-0.4562121,0.024141207,0.19423743,0.7284635,0.27046913,0.091427475,-0.27290818,2.4457474,0.35267624,-0.14124544,-0.3587356,0.6354651,0.44762415,0.5347106,0.1576927,-0.07035433,-0.62669516,-0.5964982,0.022721335,0.8997787,-0.088240206,-0.22340164,0.113081306,-0.093459226,-0.3731728,-1.0652199,0.15207438,-0.08235215,-0.38412276,-0.12380577,0.16916335,-0.37506893,0.9809917,0.505852,-0.27879626,-0.1356077,-0.87216944,1.4884256,-0.089887686,-0.35735047,-0.18952346,-0.058308598,-0.11398005,-0.80062836,0.46958363,-0.010470718,0.21347134,-0.19876844,0.04650094,-0.75897115,0.24202266,0.17335476,0.5694237,-0.1755247,0.046171527,0.4265713,0.5958072,0.27597517,-0.091395356,-0.13192715,-0.6837542,-0.10110636],"result":{"type":"object","properties":{"address":{"nullable":true,"type":"string"},"seed":{"nullable":true,"type":"string"},"walletId":{"nullable":true,"type":"string"}},"required":[]},"sql_tables":null,"sql_queries":null,"file_inbox":null,"oauth":null,"assets":null},false]}
