+++ tools/coinbase-get-balance/.tool-dump.test.json
@@ -1,0 +1,1 @@
+{"type":"Deno","content":[{"toolkit_name":"coinbase-get-balance","name":"Coinbase Balance Getter","author":"Shinkai","version":"1.0.0","js_code":"import { Coinbase, CoinbaseOptions } from 'npm:@coinbase/coinbase-sdk@0.0.16';\n\ntype WalletBalances = { [key: string]: number };\n\ntype Configurations = {\n  name: string;\n  privateKey: string;\n  walletId?: string;\n  useServerSigner?: string;\n};\ntype Parameters = {\n  walletId?: string;\n};\ntype Result = {\n  message: string;\n  balances: WalletBalances | null;\n};\nexport type Run<C extends Record<string, any>, I extends Record<string, any>, R extends Record<string, any>> = (config: C, inputs: I) => Promise<R>;\n\nexport const run: Run<Configurations, Parameters, Result> = async (\n  configurations,\n  params,\n): Promise<Result> => {\n  const coinbaseOptions: CoinbaseOptions = {\n    apiKeyName: configurations.name,\n    privateKey: configurations.privateKey,\n    useServerSigner: configurations.useServerSigner === 'true',\n    debugging: true,\n  };\n  const coinbase = new Coinbase(coinbaseOptions);\n  const user = await coinbase.getDefaultUser();\n\n  // Prioritize walletId from Params over Config\n  const walletId = params.walletId || configurations.walletId;\n\n  // Throw an error if walletId is not defined\n  if (!walletId) {\n    throw new Error('walletId must be defined in either params or config');\n  }\n\n  const wallet = await user.getWallet(walletId);\n  console.log(`Wallet retrieved: `, wallet.toString());\n\n  // Retrieve the list of balances for the wallet\n  const balances = await wallet.listBalances();\n  console.log(`Balances: `, balances);\n\n  // Convert balances to WalletBalances\n  const balanceMap: WalletBalances = {};\n  for (const [currency, amount] of balances) {\n    balanceMap[currency] = amount.toNumber();\n  }\n\n  return {\n    message: `Balances: ${JSON.stringify(balanceMap)}`,\n    balances: balanceMap,\n  };\n};\n","tools":[],"config":[{"BasicConfig":{"key_name":"name","description":"The name of the Coinbase wallet","required":true,"type":null,"key_value":null}},{"BasicConfig":{"key_name":"privateKey","description":"The private key of the Coinbase wallet","required":true,"type":null,"key_value":null}},{"BasicConfig":{"key_name":"walletId","description":"Optional wallet ID for specific wallet selection","required":false,"type":null,"key_value":null}},{"BasicConfig":{"key_name":"useServerSigner","description":"Optional server signer configuration","required":false,"type":null,"key_value":null}}],"description":"Tool for getting the balance of a Coinbase wallet after restoring it","keywords":["coinbase","balance","shinkai"],"input_args":{"type":"object","properties":{"walletId":{"type":"string","description":"Optional wallet ID to get balance for a specific wallet"}},"required":[]},"output_arg":{"json":""},"activated":false,"embedding":[-0.1122051,0.21792355,-0.019444812,-0.32676318,0.017230723,-0.042694233,0.12454068,0.23506877,-0.049858324,-0.39198625,-0.2356008,0.14091831,0.025426041,-0.14607342,0.2663238,-0.17875119,0.13934344,-0.46660352,-1.2717104,-0.044658728,0.22778314,0.56215423,-0.6181404,0.37758034,-0.56418365,-0.14707409,-0.48298508,-0.003713008,-1.6355104,-2.3401878,0.020875055,1.0987378,-0.068168,-0.25705624,-0.6406843,-0.05842889,0.046699405,0.30345416,-0.596434,-0.41274422,0.08122185,0.43440035,-0.5265702,0.0362359,-0.0654276,0.017466605,0.5859141,-0.49892268,-0.07515541,0.6399467,0.13092819,-0.02034505,0.58803767,-0.49216923,0.0382146,0.038736127,0.13882983,-0.6310905,0.26340827,0.32647747,0.024176609,0.10194284,-4.3208013,-0.040059604,0.69316846,-0.21621847,0.4555591,0.21492131,-0.28525597,0.19751486,-0.4944448,-0.21758324,-0.2561371,0.35391644,0.4861079,0.069663204,0.7510352,-0.108434245,0.4958588,-0.23860747,-0.17986467,0.8263099,0.31614143,0.0266728,-0.4415525,0.32278436,-0.059656236,-0.72129464,0.35087603,0.118561104,0.072357774,0.25650117,0.0114029525,-0.050457455,-0.36058816,-0.029316857,0.05345415,0.91371477,0.099857435,3.0752783,0.68999904,0.4103311,0.50830394,-0.4709496,0.8439158,-0.7339027,0.27937278,-0.42391467,0.64193827,-0.010285383,0.6696393,-0.16732915,0.38181823,0.46825826,0.44639584,-0.04503636,-0.7767281,-0.027618747,-0.45242998,0.5186432,-0.60994107,0.1560903,-0.70730853,-0.086867385,-0.33560067,-0.31328416,-0.4761167,0.35943118,-0.16717035,0.28757736,0.7505316,-0.4674905,-0.94920105,-0.24537762,-0.22457483,0.68390805,-0.055820018,-0.5679477,-0.5199877,-0.3298007,-0.23557621,-1.0684874,0.70712435,-0.6234385,0.17502213,-0.0074590556,0.5305823,-0.23091885,-0.26135948,-0.26235065,-0.42848885,0.3799429,0.3621992,0.20847175,0.74951464,-0.04785835,0.103263736,-0.074650764,-0.48663232,-0.04685007,0.003893502,-0.21457385,0.5851617,0.7577644,0.9477838,-0.39980376,0.5123923,0.17587909,0.40257978,-0.6337289,0.48085725,0.22692579,-0.1308599,-0.17204627,-0.3260076,-0.0532555,-0.034859188,0.14537829,0.38741913,-0.3559643,-1.2954301,0.65634984,-0.73787755,-0.6614718,-0.16321383,0.27331,0.39565712,-0.21289161,0.7244392,0.60400033,-0.63224274,1.6804485,-0.421256,-0.06698799,-0.49296844,-0.043947622,-0.2732744,0.22410235,0.609065,-0.17217714,-0.46537006,0.043895226,0.034387194,0.47640026,-0.6910904,-0.14865017,0.26914796,0.022420472,-0.07786381,-0.211391,0.5306046,0.14311999,0.014625061,0.6163252,0.38819557,-0.27364784,0.32498214,0.05358205,1.027975,0.20983253,0.2098367,0.15997908,-0.8947806,-0.6818938,-0.6106497,0.011262834,0.34579194,0.024799298,0.21171358,0.41551903,0.41326112,0.23885198,0.30851015,0.051665213,1.3319033,0.27679336,-0.063311286,0.10576472,-0.008384563,-0.8188471,0.11795208,0.05702358,-0.69708276,-0.1623407,-0.25308907,-0.5819838,0.5912054,-0.34703276,0.28800455,1.8632154,0.52673703,-0.31558284,0.41940686,0.034290798,0.10077952,0.35029146,-1.1312464,0.110920474,-0.52502084,0.26309755,0.07871712,-0.1673185,-0.2582632,0.16581045,0.07132866,-0.4274863,-0.77864736,0.62542224,0.29736987,-0.48618934,0.21518765,1.3097459,0.804146,-0.019944578,-0.20996451,-0.25136352,0.3375256,0.33029765,-0.31664985,0.23771079,0.5451549,-0.15576157,0.52026814,0.15959534,0.397811,-0.08771992,-0.30593848,-0.015177786,-0.9437044,-0.13004494,-0.0895731,-0.102244705,-0.58599967,0.7524687,1.7816204,0.73901916,0.8541392,0.30481288,-0.05781647,0.6458275,-0.3604158,0.74742454,-0.40308708,0.39093086,-1.0534573,-0.5323918,0.4550317,-0.63979226,-0.043461435,-0.24512357,-0.037757806,0.47987717,0.23262349,0.48169845,0.14022627,-0.667673,-0.7440694,0.26770002,-0.055000126,-2.257101,0.15032548,-0.59528404,-0.44336444,-0.0726046,-0.17490539,0.35630667,0.1335873,0.29353437,-0.25075904,1.7107471,0.16393793,-0.32271257,-0.70558083,0.018164277,0.875983,0.59886837,0.23601161,0.06945345,-0.7084349,0.008139738,-0.13749018,0.7753262,0.03842848,0.14583483,-0.14556053,0.008987986,-0.33083373,-0.889037,0.27736363,-0.32143906,-0.058217615,-0.044049975,0.2679112,-0.5327888,1.3052187,0.3804539,-0.33331263,0.07149097,-0.3488194,1.4735043,-0.11365883,-0.19220917,-0.47264743,-0.23412628,-0.087442644,-0.09439272,0.17180006,-0.43823156,-0.20659213,0.3114082,-0.102425635,-0.34236217,0.1895842,0.28456134,0.6217086,0.6311034,0.21342117,0.42619845,0.3926774,-0.1294566,0.02903571,-0.30436456,-0.9850829,0.3401345],"result":{"type":"object","properties":{"balances":{"additionalProperties":{"description":"Balance amount for a specific token","type":"number"},"description":"Map of token symbols to their respective balances","required":[],"type":"object"},"message":{"description":"Status message about the balance retrieval operation","type":"string"}},"required":["message","balances"]},"sql_tables":null,"sql_queries":null,"file_inbox":null,"oauth":null,"assets":null},false]}
