+++ tools/coinbase-get-my-address/.tool-dump.test.json
@@ -1,0 +1,1 @@
+{"type":"Deno","content":[{"toolkit_name":"coinbase-get-my-address","name":"Coinbase My Address Getter","author":"Shinkai","version":"1.0.0","js_code":"import { Coinbase, CoinbaseOptions } from 'npm:@coinbase/coinbase-sdk@0.0.16';\n\ntype Configurations = {\n  name: string;\n  privateKey: string;\n  walletId?: string;\n  useServerSigner?: string;\n};\ntype Parameters = {\n  walletId?: string;\n};\ntype Result = {\n  address: string;\n};\nexport type Run<C extends Record<string, any>, I extends Record<string, any>, R extends Record<string, any>> = (config: C, inputs: I) => Promise<R>;\n\nexport const run: Run<Configurations, Parameters, Result> = async (\n  configurations: Configurations,\n  params: Parameters,\n): Promise<Result> => {\n  const coinbaseOptions: CoinbaseOptions = {\n    apiKeyName: configurations.name,\n    privateKey: configurations.privateKey,\n    useServerSigner: configurations.useServerSigner === 'true',\n  };\n  const coinbase = new Coinbase(coinbaseOptions);\n  const user = await coinbase.getDefaultUser();\n\n  // Prioritize walletId from Params over Config\n  const walletId = params.walletId || configurations.walletId;\n\n  // Throw an error if walletId is not defined\n  if (!walletId) {\n    throw new Error('walletId must be defined in either params or config');\n  }\n\n  const wallet = await user.getWallet(walletId);\n  console.log(`Wallet retrieved: `, wallet.toString());\n\n  // Retrieve the list of balances for the wallet\n  const address = await wallet.getDefaultAddress();\n  console.log(`Default Address: `, address);\n\n  return {\n    address: address?.getId() || '',\n  };\n};\n","tools":[],"config":[{"BasicConfig":{"key_name":"name","description":"The name of the Coinbase wallet","required":true,"type":null,"key_value":null}},{"BasicConfig":{"key_name":"privateKey","description":"The private key of the Coinbase wallet","required":true,"type":null,"key_value":null}},{"BasicConfig":{"key_name":"walletId","description":"Optional wallet ID for specific wallet selection","required":false,"type":null,"key_value":null}},{"BasicConfig":{"key_name":"useServerSigner","description":"Optional server signer configuration","required":false,"type":null,"key_value":null}}],"description":"Tool for getting the default address of a Coinbase wallet","keywords":["coinbase","address","shinkai"],"input_args":{"type":"object","properties":{"walletId":{"type":"string","description":"The ID of the Coinbase wallet to get the address from"}},"required":[]},"output_arg":{"json":""},"activated":false,"embedding":[-0.17932753,0.15964293,-0.41342285,-0.35331523,-0.10128805,-0.103858195,0.0112680495,0.82166934,0.03933646,-0.08853404,-0.3104675,0.008514911,0.17749289,0.06472419,0.15668456,-0.57975745,0.119621426,-0.7146136,-0.67832845,0.34949583,0.24846318,1.2754143,-0.4169299,-0.045812353,-0.2568414,-0.29605535,-0.22251318,-0.39910167,-1.4526182,-2.3240116,0.22368193,1.2659607,-0.10700137,0.04635729,-0.7943327,0.056435622,0.11881925,0.37792012,-0.5881921,-0.67314845,0.22995919,0.22469546,-0.25196442,-0.2977796,-0.2808156,0.16371796,0.5224911,-0.17036061,0.08240048,0.55401254,0.5550245,-0.06726978,0.37150687,-1.0110029,-0.19860011,0.3690119,0.062547445,-0.32751927,-0.13983053,0.65957195,-0.11019101,0.28199273,-4.1167026,-0.010445392,1.1536654,-0.044021882,-0.19383013,-0.20445186,-0.45228192,0.3846407,-0.28967196,-0.3215696,-0.50890607,0.55178374,0.6466307,0.50096804,0.40960896,-0.015315123,0.9545287,-0.082977295,-0.013413724,0.64457166,0.17772853,0.32607895,0.028716639,0.20834231,-0.280965,0.022425786,0.06716788,0.67692375,-0.039648425,-0.1904221,-0.42792153,0.3081607,-0.34659982,-0.07910152,-0.13672504,0.87124807,0.25957933,2.7408013,0.82253426,0.33271676,0.3830495,-1.0432297,1.0111794,-0.3639133,-0.27953902,-0.38907802,0.5439809,-0.1665577,0.2764875,-0.3984063,0.5602261,0.39349622,0.06076364,0.14092079,-0.33769405,-0.4078202,-0.24881178,0.5352551,-0.41751012,-0.06376307,-0.91029376,-0.3063765,-0.229336,-0.03493604,0.036654722,0.3033824,-0.12646031,0.16920257,0.8081108,-0.68709177,-0.87597513,-0.20345353,0.05654283,0.5672903,0.25627428,-0.47784224,-0.48938382,-0.23010993,0.033031926,-1.432347,0.8707811,-0.11786066,0.53551245,-0.18216951,0.56844187,-0.45306367,-0.871271,0.06953817,-0.01883347,0.10133599,0.8850505,0.3210193,0.5014882,-0.68166435,0.067953594,-0.26241648,-0.09095915,0.04191948,-0.085172355,-0.29178366,0.49644402,1.0346931,0.49191988,-0.21278837,0.1552579,0.42457995,0.37517163,-0.6189191,0.65872675,0.2103756,0.4102558,0.28160647,-0.45155466,-0.2631366,-0.13053115,-0.17375982,0.34568825,-0.3599729,-1.2549177,0.38902745,-0.564535,-0.63555527,0.17842765,-0.086037025,0.6811883,0.0031095045,0.45187286,0.16761576,-0.6281824,1.3175262,-0.2932405,-0.11214225,-0.59985596,-0.12824976,-0.2609207,0.34946337,0.73185855,0.10807916,-0.6177121,-0.3831613,0.05921615,0.53959525,-0.6342268,-0.07955518,0.3043292,0.6878448,-0.46822035,0.06468081,0.76058364,-0.7113683,0.29584336,0.44136515,0.69938636,-0.22730197,0.3251933,0.11502963,0.77010596,-0.032495484,0.1063212,-0.19553521,-1.0336113,-0.5114762,-1.0078665,-0.18042813,0.49700218,0.45784438,-0.23957342,0.4018757,-0.06607162,0.15405571,0.6597649,0.2194557,0.95349145,0.806748,-0.1722436,-0.027636163,-0.032721918,-0.8663484,0.033776123,0.081975594,-0.49234384,-0.37394935,0.35717422,-0.77462137,0.7966627,-0.15514413,0.08736948,1.6222768,0.55413777,-0.029605638,-0.007258404,0.22146,0.1714007,-0.025337152,-1.3562571,0.0052703926,-0.028748669,0.99206585,0.04683815,-0.0476064,-0.037525747,0.4543084,0.00021454127,-0.5230759,-0.7134363,-0.0066314265,0.24200809,-0.4915258,0.03015158,1.1206086,0.92073715,-0.08053277,-0.5278128,-0.32479218,0.17775685,0.36103645,-0.32219997,0.053983346,0.2498126,-0.120526075,0.90176004,0.105171755,0.18076822,-0.14823234,-0.035519555,0.15436316,-1.0244967,-0.06540312,-0.09376261,-0.6566599,-0.28878072,1.0337374,1.6364617,0.88323754,0.49055243,0.27557373,0.022975512,0.92566395,-0.13089591,0.46502104,-0.4719314,0.30416837,-0.93894005,-0.5786668,0.47886258,-0.5090064,-0.3169286,-0.115220055,0.22271526,0.7059446,0.26687512,0.37598094,-0.2801255,-0.5246223,-0.7853387,0.13136622,-0.29093853,-2.1983619,-0.061054446,-0.6261363,-0.60235673,-0.32901666,-0.36539868,0.47864866,0.0012961347,0.29893708,-0.119530365,1.6799016,0.31920388,-0.53568,-0.44493467,0.2224508,0.7926569,0.6158743,0.31482908,0.18538722,-0.78197426,-0.07565202,0.048063297,1.2009664,0.1422099,0.052135646,-0.17165433,-0.039600164,-0.4821769,-1.4015799,-0.07014266,-0.11032796,-0.20598231,-0.18786888,-0.19679819,-0.29310876,1.2269748,0.7410622,-0.7246054,-0.43472102,-0.3305997,1.2603365,0.12693687,-0.09097959,-0.48109823,-0.9074406,0.096925765,0.067318454,0.22055043,0.2542562,0.24284221,-0.23175079,0.007990668,-0.15167749,0.2416907,0.010765458,0.41447288,0.30766064,0.110257745,0.2081362,0.66511655,-0.21753985,0.26317722,-0.09763156,-0.93785083,0.6041742],"result":{"type":"object","properties":{"address":{"description":"The Ethereum address of the Coinbase wallet","type":"string"}},"required":["address"]},"sql_tables":null,"sql_queries":null,"file_inbox":null,"oauth":null,"assets":null},false]}
