+++ tools/file-read/.tool-dump.test.json
@@ -1,0 +1,1 @@
+{"type":"Deno","content":[{"toolkit_name":"file-reader","name":"Read File Contents","author":"Shinkai","version":"1.0.0","js_code":"type CONFIG = {};\ntype INPUTS = { path: string };\ntype OUTPUT = { content: string };\n\nexport async function run(config: CONFIG, inputs: INPUTS): Promise<OUTPUT> {\n    const { path } = inputs;\n    const content = await Deno.readTextFile(path);\n    return { content };\n}","tools":[],"config":[],"description":"Reads the text contents of a file from the given path.","keywords":["file","read","content","text"],"input_args":{"type":"object","properties":{"path":{"type":"string","description":"The path of the file to read."}},"required":["path"]},"output_arg":{"json":""},"activated":false,"embedding":[0.12859492,0.12596734,0.010835137,-0.15759677,0.08992459,-0.06392758,-1.4077084,0.8902618,0.32812726,-0.09717176,-0.040325865,0.93691605,0.007549899,-0.51253396,0.2567432,-0.4170107,-0.3970057,-0.58185434,-1.0538585,-0.28348687,0.15490325,0.8829895,-0.014019832,-0.17916541,0.027874865,0.046194345,-0.39674753,-0.4991902,-1.0936495,-2.2655258,0.28712308,0.3062436,-0.07864686,0.2525275,0.10321572,-0.086163834,0.26947227,-0.17715436,-0.6861206,-0.3886912,-0.05326306,0.058443986,-0.5777914,0.15437311,0.19049811,0.25289434,0.1809145,-0.9324337,1.3157676,1.1605527,-0.77865005,-0.54476464,-0.25389823,-0.7968005,-0.18135706,-0.32752505,0.47686017,-0.06893139,-0.3669963,-0.3656043,-0.24349539,-0.038294002,-3.8615422,-0.21388562,0.549955,0.23996615,0.27268782,0.010611329,0.2233977,-0.54284453,-0.38535947,0.1361036,0.2554217,0.299286,0.30582452,-0.96808624,-0.047579106,-0.008844443,0.53406525,-0.40323326,-0.4771782,1.0571189,-0.25730595,0.22049019,-0.8480378,0.62866956,-0.3508037,-0.57431704,0.35690808,0.058529556,0.41120374,-0.37303326,0.22492631,0.27541775,0.021659356,-0.32826716,-0.19713931,0.60175943,0.56584424,2.7791252,1.0450399,0.594266,0.20650783,-0.61168516,0.2008933,-0.4747657,0.024587486,-0.01762429,0.010493876,-0.62326825,0.3704437,-0.47063246,-0.14854854,0.24279964,-0.12406143,-0.30222398,-0.48557103,-0.08278459,0.63487,0.67132777,-0.99932593,-0.08010429,-0.6655833,-0.4096786,0.06130149,-0.34929386,0.55783796,0.19386198,0.19404489,-0.33255413,0.8587968,-0.2075023,-0.66796803,0.13709542,-0.060420133,0.4835882,0.3852095,-0.9448407,0.36258578,-0.66567135,-0.15313038,-1.3003845,0.69442594,0.19294913,0.38094082,0.13808331,-0.10014426,0.30711347,-0.35933435,0.12892267,0.42373046,0.15688801,-0.09071456,0.23043248,0.42913413,0.41824204,-0.18021575,0.11129397,-0.33085412,0.32014716,-0.71196103,-0.17633112,-0.19455895,0.5147253,0.32320574,0.042989697,-0.048838228,0.1935009,0.16269538,0.032192305,1.1123198,-0.84431124,-0.67001206,0.29565674,-0.4992154,-0.020182084,-0.49168688,0.3178972,-0.20264104,-0.036908656,0.15182772,0.4023808,-0.6536689,-0.15701845,-0.41465968,0.43057722,0.6848841,0.30100897,1.6196972,1.1918077,-0.37824315,2.4926894,-1.3363681,-0.59464365,0.027451769,0.08557453,0.03887879,-0.026824325,-0.10933608,-0.11641621,0.7005088,-0.011468995,0.3816275,-0.16541559,-0.20133889,-0.7128427,0.61280257,0.035759382,-0.24480699,-1.11442,0.49035117,-0.2760281,0.190466,0.5364733,0.6639886,0.25680214,-0.31205228,0.24600632,-0.10684079,0.64871716,-0.7363786,0.4419547,0.17056711,-0.36390266,-0.6615776,0.9569794,0.53285533,0.3697747,-0.3690424,-0.5845392,0.0757478,0.8685907,1.551799,0.77571744,0.7099622,0.62299573,-0.047043465,0.6500958,-0.010198675,-0.22083391,-0.090329014,-0.027690284,0.2020743,-0.34220135,-0.13239811,-0.74922097,0.34367022,0.4056856,-0.2654303,1.3016803,0.4247997,0.26406908,-0.27320027,0.31540078,0.32827455,0.12793702,-1.9450389,-0.38419828,0.075002335,0.15604155,0.1467857,-0.45322528,0.2651813,0.24101184,0.4704963,-0.21165113,-0.77044237,-0.7644199,0.06674867,0.29760006,-0.6619981,0.43155158,-0.18134195,-0.2950378,0.08561646,-0.21715693,0.4669162,0.5292284,-0.24674854,0.06567959,0.28479904,0.02634546,0.3878179,-0.45681316,-0.15824303,0.08249477,-0.085056454,-0.31129003,0.1623993,0.36038527,-0.35157487,-1.3610289,-0.7839526,0.30959064,1.7284135,-0.094772324,0.11379894,0.41203946,0.045135368,0.20834355,0.3884402,-0.104922004,-0.6168906,1.0942903,-0.048844784,-0.18529612,-0.0014815256,0.05581587,0.04074034,0.26128578,-0.84257406,-0.028878443,0.16798304,-0.14255713,0.3751877,-0.15724911,0.47445253,-0.20793211,-0.42448515,-1.8821567,-0.46138513,-0.05275587,0.19684908,-0.0017875135,-0.3411157,0.9906085,-0.6099685,-0.10422182,-0.2852644,1.0378195,0.73284316,-0.25559312,-0.38850874,-0.11635728,0.4252638,0.52667606,-0.056361556,-0.08761079,-0.22849727,0.07162489,0.5779581,1.3472354,0.24429214,0.7627983,0.8608464,0.11962003,-0.52168983,-1.3894522,-0.21532476,-0.41268677,-0.51807946,0.5023647,0.1996884,-0.17716783,-0.35021126,0.61440194,0.0011843555,0.10789876,-1.0063789,2.3573909,-0.23840012,0.39342797,-0.3351542,-0.12421349,0.1322378,-0.22049274,0.42365342,-0.29355833,-0.25301337,0.03335283,0.35332134,-0.1319727,-0.017306782,0.2564673,0.88989186,0.58354783,0.16554339,0.42174625,-0.111497805,0.37349415,0.51318854,0.077761844,-0.7079764,-0.31325018],"result":{"type":"object","properties":{"content":{"description":"The content of the file.","type":"string"}},"required":["content"]},"sql_tables":[],"sql_queries":[],"file_inbox":null,"oauth":null,"assets":null},false]}
