+++ tools/file-read/metadata.json
@@ -1,4 +1,5 @@
 {
+    "id": "file-reader",
     "name": "Read File Contents",
     "version": "1.0.0",
     "description": "Reads the text contents of a file from the given path.",
