+++ tools/file-update/.tool-dump.test.json
@@ -1,0 +1,1 @@
+{"type":"Deno","content":[{"toolkit_name":"file-update","name":"Update File with Prompt","author":"Shinkai","version":"1.0.0","js_code":"import { shinkaiLlmPromptProcessor } from './shinkai-local-tools.ts';\n\ntype CONFIG = {};\ntype INPUTS = { path: string, prompt: string };\ntype OUTPUT = { new_file_content: string, message: string };\n\nexport async function run(config: CONFIG, inputs: INPUTS): Promise<OUTPUT> {\n    const file_content = await Deno.readTextFile(inputs.path);\n    const prompt = `\n<${inputs.path}>\n${file_content}\n</${inputs.path}>\n\n<rules>\n * Only respond with the new file content, from the start of the file to the end.\n * Do not include any other text or comments.\n * Apply the following instructions in the tag to \"${inputs.path}\"\n</rules>\n\n<instructions>\n    ${inputs.prompt}\n </instructions>\n`;\n    const update = await shinkaiLlmPromptProcessor({ prompt });\n    await Deno.writeTextFile(inputs.path, update.message);\n\n    return { \n        new_file_content: update.message, \n        message: \"File updated\"\n    };               \n}","tools":["local:::rust_toolkit:::shinkai_llm_prompt_processor"],"config":[],"description":"Applies a prompt to the file contents.","keywords":["file","update","prompt","text"],"input_args":{"type":"object","properties":{"path":{"type":"string","description":"The path of the file to update."},"prompt":{"type":"string","description":"The prompt to apply to the file contents."}},"required":["path","prompt"]},"output_arg":{"json":""},"activated":false,"embedding":[-0.039903395,0.1478697,-0.37138444,-0.2536462,0.42034775,0.19308072,-0.73079115,0.79844755,0.5390755,-0.2896065,-0.162473,0.70904064,-0.32502642,-0.18482137,-0.09624794,-0.2651251,-0.609646,-0.81935316,-0.88271856,-0.01937908,-0.10085076,0.55340976,0.37698966,0.39858586,0.53191835,0.6502075,0.09632544,0.211497,-1.5360123,-2.6435096,0.19663453,0.44426164,0.23009388,0.0703166,0.9099579,-0.48099,-0.036563464,0.00794445,-0.46558335,-0.39092416,0.610461,0.19468907,0.038887687,-0.3964421,-0.10741798,-0.14740476,0.40959883,-0.53380513,1.2657025,0.2965763,-0.9813956,-0.43001893,-0.54247403,-0.007885814,0.09385339,-0.0095655285,-0.022299696,-0.16926739,-0.43881342,0.21310613,-0.5904309,0.44975644,-3.9330294,0.003366638,0.2304686,0.41677976,-0.40793145,0.24047405,0.2959547,-0.55440354,-0.59958905,0.31358656,-0.42083263,0.5289503,0.15836212,-1.471573,-0.09732093,-0.021677233,0.3981617,-0.58984214,-0.06760834,0.75852627,-0.30363005,-0.4435111,-0.82126266,0.62332565,-0.19124517,0.17386177,0.24707484,0.16973849,0.2766788,0.36825925,0.68306947,0.034848794,0.20050102,-0.08022639,0.28803945,-0.08378753,0.15303749,2.8060524,0.46759516,0.47084996,0.59963673,-0.14851823,0.3826089,-0.14490002,-0.30825695,-0.49160218,-0.26169422,-0.5537072,0.29882553,-0.41973025,-0.33127865,-0.08648269,0.64579785,-0.38195258,-1.1662104,0.14285554,-0.08915171,1.1903292,-0.2657309,-0.29888767,-0.8791976,-0.052203894,-0.24990591,0.25232184,0.39312047,0.112233624,0.20348947,-0.12953644,0.53399,-0.04749395,-0.5620714,-0.7673449,0.35613254,0.6470855,0.6655971,-0.33039773,0.49075794,-1.0007432,0.07129872,-0.8365802,0.83558166,-0.24609178,0.30504823,-0.2184625,-0.7150377,-0.03675293,-0.4239329,-0.83338344,-0.30540553,0.31882566,0.44099975,-0.033006478,0.44189107,0.1351082,-0.017860558,0.4303109,-0.4026429,0.46499673,0.23424427,0.021554373,0.24941018,0.6592791,0.53289306,-0.52218735,0.06256876,0.011257041,-0.42429513,-0.0031392332,0.02167125,-0.6984439,-0.039923962,0.14456223,-0.4790928,0.38158292,-0.09277673,0.1220977,0.50893325,-0.43329892,0.052438244,0.4842253,-1.397022,-0.4553201,-0.10356268,0.3240641,0.067737706,0.18493679,0.5851721,1.0497124,-1.0224062,1.8358845,-1.4865811,-0.4442987,-0.2588486,0.46076733,-0.052975863,0.39531308,-0.42411664,0.19004482,0.28026456,0.05875508,0.47741306,-0.038415417,-0.32223785,-0.7042106,0.5208464,0.8885053,-0.40057737,-0.3352108,-0.03342961,-0.07285142,0.759855,0.27799243,1.1157476,0.15175332,-0.13291155,0.29058608,-0.16938427,0.8717817,-0.41137427,0.8388778,-0.47184545,-0.2556157,-0.5926424,0.2109636,-0.10387155,-0.029345576,-0.2337321,-0.36045682,0.6727238,1.0322102,0.85615,0.4178601,0.060016125,0.4152884,-0.18892214,0.48241818,0.20748389,-0.42650983,-0.42875284,-0.1383827,0.416398,-0.26945934,0.26700795,-0.8047602,-0.045017943,0.038441703,-0.15906563,1.1843193,0.6012529,0.088855594,0.55948305,-0.2843043,0.06701256,-0.0033614188,-1.5874362,-0.48762953,-0.26017806,0.096587494,0.22007397,-0.31763422,0.28027433,0.393404,0.09217901,-0.6059741,-0.4499374,-0.041973565,0.34829432,0.3226861,0.01581041,0.20072733,0.047392674,-0.26573905,-0.23842724,-0.06447384,0.51573956,0.5406328,-0.31117725,-0.45689532,-0.41448826,-0.2663731,-0.49987212,-0.44064802,-0.19571939,-0.51080966,-0.27360648,-0.19680703,0.00872729,0.8080794,-0.19187385,-0.49985233,-0.64229125,0.50071347,2.379929,0.045338828,0.6211416,0.85915077,-0.3084452,0.05690112,0.42632666,0.10319888,-0.34689292,0.5453186,-0.5494945,-0.029376745,-0.27340746,-0.105934694,0.118975565,0.64662135,0.14513269,-0.08858763,-0.5285297,-0.2954827,0.5633896,-0.47422564,0.49663198,0.5022613,-0.036011487,-1.8981999,-0.7474049,-0.14800891,-0.23608637,0.03129068,-0.28215757,0.5260432,0.44269934,0.8593093,0.0017987639,1.2411449,0.61670935,-0.5042192,-0.27931187,-0.09320265,0.6017174,-0.17417188,-0.021029899,-0.063922755,0.08209244,-0.19690697,1.0260704,1.8778706,0.28716427,0.5454953,0.22317907,0.50895196,-0.21493927,-1.5025272,-0.199907,0.17876531,-0.73794264,0.6096667,-0.11621555,-0.4040492,-0.15010911,0.5291127,0.43814874,-0.14281955,-0.5761751,1.7517116,-0.037796408,-0.5165784,-0.2904388,0.28494275,0.40139303,-0.2760628,0.3002559,-0.3823212,-0.9122817,-0.284226,0.08467552,-0.14658695,-0.17697525,0.40925556,0.39127195,0.8325679,0.11660272,0.49203834,0.35033125,0.31263828,0.20281029,0.52836424,-0.66714835,-0.07887113],"result":{"type":"object","properties":{"message":{"description":"The message returned from the tool.","type":"string"},"new_file_content":{"description":"The path of the file that was updated.","type":"string"}},"required":["message"]},"sql_tables":[],"sql_queries":[],"file_inbox":null,"oauth":null,"assets":null},false]}
