+++ tools/file-update/metadata.json
@@ -1,4 +1,5 @@
 {
+    "id": "file-update",
     "name": "Update File with Prompt",
     "version": "1.0.0",
     "description": "Applies a prompt to the file contents.",
