+++ tools/file-write/.tool-dump.test.json
@@ -1,0 +1,1 @@
+{"type":"Deno","content":[{"toolkit_name":"file-writer","name":"Write File Contents","author":"Shinkai","version":"1.0.0","js_code":"\ntype CONFIG = {};\ntype INPUTS = { path: string, content: string };\ntype OUTPUT = { message: string };\n\nexport async function run(config: CONFIG, inputs: INPUTS): Promise<OUTPUT> {\n    try {\n        await Deno.writeTextFile(inputs.path, inputs.content);\n        return {\n            message: \"File written\"\n        };               \n    } catch (error) {\n        return {\n            message: \"Failed to write file \" + error.message,\n        };               \n    }\n}","tools":[],"config":[],"description":"Writes the text contents of a file to the given path.","keywords":["file","write","save","content","text"],"input_args":{"type":"object","properties":{"path":{"type":"string","description":"The path of the file to write to."},"content":{"type":"string","description":"The content to write to the file."}},"required":["path","content"]},"output_arg":{"json":""},"activated":false,"embedding":[-0.014549136,0.3306045,0.010623809,0.34710795,0.17257398,0.06644237,-1.4433259,0.79340667,0.3200154,0.14200212,-0.013457694,1.1742105,-0.05998533,-0.27217516,0.30608752,-0.53545934,-1.1113505,-0.4852112,-1.28673,-0.52062255,-0.045713667,1.0082848,0.1051455,-0.280425,0.4606821,0.040276006,0.4114635,-0.24108294,-1.0731835,-2.4765906,0.08310933,0.45401487,-0.21076798,-0.006898947,0.45140323,-0.095220305,0.21838978,0.029122759,-0.42895025,-0.44942367,0.06679413,-0.24396394,-0.2146438,-0.19280109,-0.10422532,0.48197737,0.13546635,-0.6760576,1.1726547,0.8601404,-0.66012675,-0.29620454,-0.8183398,-0.72856545,-0.24246916,-0.27076128,0.03431056,-0.031293496,-0.15109776,-0.017892722,-0.20798804,0.39865324,-3.9953966,0.0006713234,0.7178922,0.18328021,0.33092046,0.3336682,-0.26151973,-0.32545686,-0.49092036,0.40152624,-0.052238654,0.87046874,0.54039514,-1.1568112,-0.0207678,-0.18181716,0.57120806,-0.69160694,-0.17649092,0.9864096,-0.22806545,0.5767491,-0.8603246,0.6574369,0.014650822,-0.10360512,0.3105017,-0.2222869,-0.06558968,-0.42926064,0.20834538,0.011756942,-0.22573015,-0.13948719,-0.38368568,0.5166678,0.7128568,2.7727962,0.7776012,0.22696115,0.46952018,-0.8915465,-0.033291057,-0.45063156,-0.35056576,0.042245828,-0.4374227,-0.89509064,0.3355586,-0.24818873,0.039706945,-0.32526097,0.5800467,0.08923647,-0.52462393,0.10765308,0.23816544,0.7645004,-0.78018534,-0.070886716,-0.8608689,-0.08300616,-0.22955745,0.19226775,0.5603884,0.16824265,0.25708127,-0.33594754,0.51311,-0.33226657,-0.15299347,0.34936494,-0.2556605,0.6391561,0.13773233,-0.92406696,0.54077995,-0.22816184,0.8931076,-1.1649122,0.99441516,0.37736514,0.5239982,-0.038703598,-0.5050386,-0.110050656,-0.6058962,-0.21717086,-0.16982521,0.28685325,-0.28068703,0.118647784,0.704295,-0.10715413,-0.09661652,0.3075416,-0.35882607,0.2968553,-0.3788554,-0.1638914,-0.07119171,0.75048673,0.18922323,-0.092549145,-0.08562735,-0.10826003,-0.14149918,0.023589768,0.6068331,-0.9169399,-0.449064,0.22993019,-0.7784011,0.14004722,-0.19789673,0.4141601,0.07210271,0.43419883,0.58937657,0.30005506,-0.5529813,-0.4282561,0.057661355,-0.018719653,0.40104038,0.27569667,1.1413771,1.1122822,-0.8609533,2.174323,-1.3083888,-0.16884467,0.12011647,0.28775114,0.14208685,-0.04532428,-0.17684196,-0.5879013,0.8283146,0.13116159,0.05748804,0.12706596,0.23129657,-0.80002236,0.026521016,0.5577789,-0.19492598,-0.78969574,0.58088875,-0.71387744,0.5685177,0.66645646,1.0259598,0.06320852,-0.5522171,0.66180646,0.26632988,0.20901069,-0.45640618,0.43380922,0.27643025,-0.40231875,-0.5272685,0.58058095,0.55033726,0.3938542,-0.10604507,-0.92763895,0.572325,1.0389303,1.3557471,0.7059471,0.6976317,0.22795168,0.21427763,0.49157652,-0.21424726,-0.51271,0.036108606,0.17231429,0.384692,-0.46531135,-0.5072291,-0.8673363,0.27809343,0.4017111,-0.08000428,1.6100259,0.21671647,0.40271592,-0.22550207,-0.081755266,0.45057532,0.55846816,-2.002303,-0.27432284,-0.6414228,0.28061464,0.3093992,-0.107880086,0.21025568,0.2429951,0.14653276,-0.6666288,-0.35030785,-0.6874968,-0.061106198,-0.013156585,-0.7392308,0.112501286,-0.02235203,0.107314825,0.29493564,-0.26220566,0.35689977,0.33182338,0.050738946,0.0029906612,0.23711354,-0.20861107,0.29840007,-0.34016418,-0.34782887,-0.2183025,0.10393673,-0.4915806,0.59145266,0.3148432,0.0968147,-1.1266612,-0.7889908,0.618552,2.0173414,0.4167422,0.4806697,0.044930957,0.05904943,-0.32249916,0.29608616,0.05759141,-0.31402722,0.74100035,-0.2828127,-0.3926125,-0.3492974,-0.25274813,0.21653771,0.40253896,-0.1648409,-0.03159745,0.035394523,-0.26036087,0.1953007,0.2856652,0.2797497,-0.22072849,-0.44991183,-1.522079,-0.008879771,-0.21769571,-0.01930108,0.008666411,-0.24933182,1.0830514,0.09082381,0.2020956,-0.056170292,0.6714194,0.566682,-0.69803435,-0.29342037,0.11870221,0.36008102,0.058972415,0.15174335,-0.18342644,-0.056950834,-0.56210285,0.69107586,1.0892509,-0.05280915,0.43005684,0.67724687,0.013258614,-0.4174863,-1.4324603,0.043559954,-0.69770515,-0.84633875,0.173196,-0.041902635,-0.08934089,-0.38253176,0.33225664,-0.08871407,0.06342561,-1.1730163,2.3047776,-0.30448312,0.43343067,0.11150242,-0.056016203,0.009526557,-0.3027591,0.5032218,-0.58914036,-0.28574032,-0.30649444,0.49732503,0.0952326,-0.005971417,0.39200407,0.9712191,0.05951012,-0.07585195,0.43632075,0.045046944,0.4728228,0.3124175,0.01903171,-0.64349025,-0.45729995],"result":{"type":"object","properties":{"message":{"description":"The message returned from the operation.","type":"string"}},"required":["message"]},"sql_tables":[],"sql_queries":[],"file_inbox":null,"oauth":null,"assets":null},false]}
