+++ tools/google-search/.tool-dump.test.json
@@ -1,0 +1,1 @@
+{"type":"Python","content":[{"toolkit_name":"google-search-shinkai","version":"1.0.0","name":"Google Search","author":"Shinkai","py_code":"# /// script\n# dependencies = [\n# \"googlesearch-python\"\n# ]\n# ///\nfrom googlesearch import search, SearchResult\nfrom typing import List\nfrom dataclasses import dataclass\nimport json\n\nclass CONFIG:\n    pass\n\nclass INPUTS:\n    query: str\n    num_results: int = 10\n\nclass OUTPUT:\n    results: List[SearchResult]\n    query: str\n\nasync def run(c: CONFIG, p: INPUTS) -> OUTPUT:\n    query = p.query\n    if not query:\n        raise ValueError(\"No search query provided\")\n\n    results = []\n    try:\n        results = search(query, num_results=p.num_results, advanced=True)\n    except Exception as e:\n        raise RuntimeError(f\"Search failed: {str(e)}\")\n\n    output = OUTPUT()\n    output.results = results\n    output.query = query\n    return output","tools":[],"config":[],"description":"This function takes a question as input and returns a comprehensive answer, along with the sources and statements used to generate the answer.","keywords":["search","answer generation","fact extraction","wikipedia","google"],"input_args":{"type":"object","properties":{"query":{"type":"string","description":"The search query to look up"}},"required":["query"]},"output_arg":{"json":""},"activated":false,"embedding":[0.74037224,0.1523448,0.12250063,-0.2330956,-0.0732149,-0.49121237,-0.7105934,0.26602554,-0.31485674,0.32427734,-0.39323238,0.7821352,0.20446597,-0.41919872,0.67798615,0.209939,0.29757926,-0.29446813,-1.6610041,-0.031176,0.78383785,0.008370239,0.25008854,-0.19770642,-0.37513775,-0.14639914,0.38888413,-0.30902037,-1.0635359,-1.5362298,0.90626204,0.7395651,-0.22855942,-0.2196196,-0.6552673,-0.38012552,-0.16014673,-0.25391504,0.018916853,-0.016465046,-0.24382688,0.05679094,-0.54244953,0.39270037,-0.2572619,-0.008627504,0.44547108,-0.057402875,0.90210915,0.88670576,-0.19960696,-0.40889594,-0.19300205,0.32754079,-0.41249725,0.20653987,-0.18449533,-0.7727064,0.06246579,-0.5266049,0.4453819,0.28019792,-4.33095,-0.010926982,0.6433091,0.18808162,0.22580339,-0.61157745,0.18193504,0.10960609,0.0867686,0.1546536,0.3066386,0.27415034,0.00094724447,-0.22532904,-0.3923833,0.025509944,0.40191758,-0.5102701,0.07003463,0.056470983,0.060049966,0.27349964,-0.38567942,0.48457092,0.1265299,-0.383452,0.36558148,-0.0033628326,-0.66018665,-0.033859618,-0.31141108,-0.13672534,-0.6319758,-0.3046636,0.28244495,0.11184969,0.42252427,3.1190393,0.2897966,-0.40488464,0.4237477,-1.1255531,0.8605757,-0.145872,0.4461098,-0.37404382,0.7425561,0.15921599,0.2937993,-0.032813367,-0.22705385,0.51221126,-0.36465305,0.20450036,-0.24083848,0.123129666,0.3802647,0.8089271,0.07132991,0.232473,-0.4389687,-0.59644294,0.057721782,0.39618132,-0.3334279,0.33608884,0.5573402,0.34885535,0.70442456,-0.39812747,-1.1715777,-0.12136582,0.13247031,0.30338153,0.3793379,-0.78292507,0.6928125,-0.5691386,0.27588916,-1.6528674,0.54117787,-0.10768123,0.59654546,-0.054497033,-0.72086287,-0.037624903,-0.413778,-0.68734956,0.10164196,0.21509701,0.3415778,0.014826264,0.7090687,0.52430075,-0.26726043,0.04955519,-0.38586935,0.17126185,-0.3468741,0.46524423,0.63281727,0.18143305,0.52894264,-0.637046,0.44291666,0.4520656,-0.19375655,-0.41542503,0.47489196,-0.4161822,-0.24842288,0.45850873,-0.04690706,-0.5877953,0.048230067,-0.14095941,0.29211515,-0.69963264,-0.33471602,0.8007372,-0.65457326,-0.4604039,-0.45024818,0.47310352,0.045648046,0.14077257,0.80203354,1.3165474,-0.57164836,1.6299654,-0.40857363,-0.60397077,-0.27461904,-0.12910345,-0.18700455,0.43255436,0.549191,0.10527582,-0.83985114,-0.12148626,0.07743675,-0.40446636,-0.16725326,-0.43537217,0.08425541,-0.123337895,-0.29343027,-0.70511866,0.28347248,-0.66800463,0.6264798,0.29570436,0.4059022,-0.080531366,-0.039702266,0.2654012,-0.7448296,0.7567029,-0.004951909,-0.038217224,-0.66041446,-0.60762405,-0.45770466,0.14596814,0.11970285,0.07908257,-0.1553611,-0.035996567,0.8584984,1.055522,-0.010617681,1.3047324,0.8748694,0.42517588,-0.10457472,0.27154553,-0.3464195,-0.35959387,0.2783715,-0.24053133,-0.3509732,-0.26509744,0.24651207,-0.65943944,0.2800332,-0.11088933,-0.009066327,1.3665054,0.9938594,0.2110039,0.21169847,0.6469496,-0.12677836,0.01716204,-1.9237192,0.019135464,-0.3490658,0.39400783,-0.22484699,-0.08604026,0.5508134,0.09623649,-0.23745012,-0.7042359,-0.075872876,-1.0612733,0.009017386,0.028394379,0.14629984,0.52394146,-0.21165091,0.11670326,0.42965937,0.23219067,-0.22103347,0.5378713,-0.7077015,-0.4177954,0.5645298,-0.027265925,-0.19830701,0.189426,-0.08153864,-0.23689003,-0.33448055,-0.046354167,0.076794356,0.20690316,-0.6722652,-0.52268887,-0.30469036,0.33876172,1.5162745,-0.03604217,0.029588792,1.017168,-0.086855166,-0.051238023,0.3323765,0.6857262,0.07683526,-0.19591287,0.0044383705,-0.77895075,0.59338504,-0.18997079,-0.26441965,0.20731027,-0.7421641,-0.09967218,0.05133872,-0.41912264,0.14718735,-0.6373423,0.49263602,0.58286464,0.29755393,-2.0216033,-0.007959459,0.40863988,-0.2722516,-0.056280687,-0.492824,0.3854516,-0.37532952,0.4678018,-0.7883893,1.9486378,0.5012205,-0.36526364,-0.46191347,0.09727245,0.674515,-0.69980514,-0.24710421,-0.10073779,-0.15110654,0.2634647,0.54191643,1.5983393,0.5812038,0.00916383,0.12389828,0.1053821,-0.377112,-1.3136234,-0.4055855,0.3349371,-0.30854625,0.43692365,-0.28489202,-0.5596402,0.46159178,1.0071299,0.531482,0.50845766,0.37580737,1.7201524,0.24967028,-0.21856175,-0.3468639,-0.7028712,-0.47778672,0.37018847,0.3560745,-0.5624895,0.423697,-0.23882222,-0.10672356,-0.33050084,0.28650072,0.29336917,0.31674817,0.65916663,0.39147034,0.81462157,0.037891425,-0.063155815,0.42848307,0.10547708,-0.5769465,-0.5225401],"result":{"type":"object","properties":{"query":{"type":"string"},"results":{"items":{"properties":{"description":{"type":"string"},"title":{"type":"string"},"url":{"type":"string"}},"required":["title","url","description"],"type":"object"},"type":"array"}},"required":["query","results"]},"sql_tables":[],"sql_queries":[],"file_inbox":null,"oauth":null,"assets":null},false]}
