+++ tools/math-exp/.tool-dump.test.json
@@ -1,0 +1,1 @@
+{"type":"Deno","content":[{"toolkit_name":"math-exp","name":"Math Expression Evaluator","author":"Shinkai","version":"1.0.0","js_code":"import { Parser } from 'npm:expr-eval@2.0.2';\n\ntype Configurations = {};\ntype Parameters = {\n  expression: string;\n};\ntype Result = { result: string };\nexport type Run<C extends Record<string, any>, I extends Record<string, any>, R extends Record<string, any>> = (config: C, inputs: I) => Promise<R>;\n\nexport const run: Run<Configurations, Parameters, Result> = (\n  _configurations,\n  params,\n): Promise<Result> => {\n  try {\n    const parser = new Parser();\n    const expr = parser.parse(params.expression);\n    const result = expr.evaluate();\n    return Promise.resolve({ ...result });\n  } catch (error: unknown) {\n    console.error('Error evaluating expression:', error);\n\n    let errorMessage =\n      'An unknown error occurred while evaluating the expression';\n\n    if (error instanceof Error) {\n      errorMessage = error.message;\n    } else if (typeof error === 'string') {\n      errorMessage = error;\n    }\n\n    throw new Error(`Failed to evaluate expression: ${errorMessage}`);\n  }\n};\n","tools":[],"config":[],"description":"Parses and evaluates mathematical expressions. It’s a safer and more math-oriented alternative to using JavaScript’s eval function for mathematical expressions.","keywords":["math","expr-eval","shinkai"],"input_args":{"type":"object","properties":{"expression":{"type":"string","description":"The mathematical expression to evaluate (e.g., '2 + 2 * 3')"}},"required":["expression"]},"output_arg":{"json":""},"activated":false,"embedding":[0.36967367,0.13820489,-0.18018691,-0.06061746,0.053862732,0.21220618,-0.22877245,-0.3107823,0.5681798,-0.10842194,-0.18641677,1.0145801,0.09758824,-0.46871328,0.44288108,-0.8003854,-0.27547854,0.029806692,-2.351401,-0.40808445,0.16066542,0.19848609,0.3527466,0.43821186,0.18010595,0.1713535,0.33997545,-0.33316314,-1.4285688,-1.6645143,0.18303299,0.5108868,-0.67435217,-0.15233773,-0.026045978,-0.21808416,0.0013305669,-0.00858137,-0.5878841,-0.6889627,-0.17357582,0.1827068,-0.4159857,-0.45987532,0.5200226,0.28148398,0.5548943,-0.2109483,0.71041495,0.17807344,-0.5695672,-0.3324992,0.51071566,0.21442385,0.098044075,-0.07878122,-0.29550236,-0.8045347,0.035856985,0.1700454,-0.29152042,0.48329982,-3.8031995,0.140291,1.0017263,0.37491223,0.21230449,0.12248981,-0.15960693,0.9140658,0.08892634,0.41345903,-0.2567477,0.13144514,0.52695376,-0.9829095,-0.087622836,0.10801749,0.8494919,-0.028191768,0.4844833,-0.123704284,-0.149269,0.49964926,-0.36645904,0.7687101,-0.17327076,-0.21860676,0.2667747,-0.32258928,-0.66398025,-0.39145353,-0.039533325,0.1479331,-0.65678895,0.36573893,0.34485385,0.108419515,0.61689764,2.9928799,0.4950359,-0.19744644,0.047998685,-1.2511368,0.43045107,-0.22639002,-0.28925052,0.17166206,0.5936987,0.032554977,0.2987402,-0.4508573,-0.3053726,0.29118514,0.27165344,0.62433386,-0.22101647,0.39120042,0.22517507,0.4003595,0.33603022,0.35778207,-0.4440238,0.1436653,-0.14749385,0.3546014,-0.7918863,0.43876657,-0.055472717,-0.017968806,0.6531599,0.03704976,-0.95477337,0.2029008,0.21212804,-0.3637649,0.4170754,-0.5314295,0.9451271,-0.06270531,0.4981871,-1.4777926,1.0789618,-0.27389845,1.0942268,0.15545376,-0.36948922,-0.06654019,-0.27083135,-0.21063285,-0.37044904,0.38381377,-0.084031776,0.29063785,1.0776656,0.6249772,-0.5797476,-0.3145232,0.25582728,0.22303936,0.08337048,0.10433025,0.3118384,0.682416,0.20981771,-0.7630354,0.42949808,0.28173724,-0.055977806,0.028277978,0.07644211,-0.1500487,-0.21876055,0.63717824,-0.24336761,-0.392952,0.7478404,-0.0045279264,0.2781953,-0.28795373,-0.07181533,0.50455433,-0.66905963,-0.49503186,0.11323649,0.32155833,0.41603893,0.36239958,0.19338244,0.48429978,-0.4667752,1.3982738,-0.80857515,-0.5337214,-0.32826,-0.05005458,-0.020430543,0.71022844,0.4500097,0.58962274,-0.47933018,-1.2188488,-0.039289623,-0.07938965,-0.27716693,-0.950839,0.51714313,-0.48108435,0.29074702,-0.42495376,-0.22212024,0.24877706,1.1656115,0.07707703,-0.021515436,0.17418396,-0.049847133,-0.21339005,-0.63622856,0.34684995,0.470005,-0.0014811605,-0.323857,-0.7038424,-0.39702195,0.65460336,-0.34012085,0.58842134,-0.60192543,-0.31443992,0.14376505,0.8322249,0.6937234,1.423618,0.2878114,-0.16883424,0.16429164,0.6032804,0.18053511,-0.04685873,0.26696566,-0.3231273,-0.6927969,0.30437207,-0.32643548,0.05459039,0.057466283,-1.0489061,-0.5959848,1.6899302,0.9979789,-0.3295823,0.13252099,0.51117814,-0.662587,-0.66738385,-2.0389743,-0.1492581,-0.8409445,0.22102407,-0.9486347,-1.1587538,0.9875838,0.76368,-0.6034394,0.079189554,-0.04153853,-0.7219739,-0.12851311,0.3494278,-0.15630601,0.28968954,-0.049051583,-0.34130186,-0.48374012,0.06986832,0.27204108,-0.039116576,-0.76302004,-0.2314945,0.28012154,0.51277715,-0.13342395,0.38889462,-0.3559021,-0.32951733,0.06259091,0.42546517,0.3081396,0.579204,-0.8814852,-0.104001544,-0.6395018,0.3358991,0.85187197,0.98876977,-0.345068,0.48206925,-0.6907707,-0.17400067,-0.8074576,0.26242837,0.17852136,-0.5406647,-0.2205798,-0.09535846,0.7594191,-0.39958072,-1.0405165,-0.24997479,-0.39850602,0.3604246,0.6563322,-0.05293411,0.7517185,-0.4574319,0.91073626,0.40672532,-0.19313775,-1.4429543,-0.48297805,0.45363176,-0.443804,-0.9323735,-0.6854489,0.4151727,-0.019103779,0.002061028,-0.46029502,1.4152046,0.6069861,0.24967723,-0.63946867,0.12329558,1.2102864,-0.6938691,0.06031894,0.04400401,0.070920385,-0.07568397,0.8633082,1.2616405,0.15856035,0.57713014,-0.12644015,0.22147396,-0.23180798,-1.1246669,-0.006440513,-0.21734972,-0.41508818,1.2805828,0.2694194,0.26632312,-0.0282857,0.5457553,0.4101187,0.12649411,-0.55052865,2.0240014,-0.13674003,-0.4983088,-0.49487492,-0.5951213,0.20916791,0.1301647,0.5942216,-0.60187054,-0.064847246,0.8937914,0.3362784,-0.39291066,0.2612105,-0.22315742,0.5770974,0.30526137,-0.18634716,0.40400478,-0.022974372,0.104909,0.12522572,-0.10052472,-0.6474921,-0.29402596],"result":{"type":"object","properties":{"result":{"description":"The evaluated result of the mathematical expression","type":"string"}},"required":["result"]},"sql_tables":null,"sql_queries":null,"file_inbox":null,"oauth":null,"assets":null},false]}
