+++ tools/math-exp/index.test.ts
@@ -1,6 +1,6 @@
 import { expect } from 'jsr:@std/expect/expect';
-import { definition } from './index.ts';
+import { run } from './tool.ts';
 
 Deno.test('exists definition', () => {
-  expect(definition).toBeInstanceOf(Object);
+  expect(run).toBeInstanceOf(Function);
 });
