+++ tools/perplexity-api/.tool-dump.test.json
@@ -1,0 +1,1 @@
+{"type":"Deno","content":[{"toolkit_name":"perplexity-api","name":"Perplexity API","author":"Shinkai","version":"1.0.0","js_code":"type Configurations = {\n  apiKey: string;\n};\ntype Parameters = {\n  query: string;\n};\n\ntype Result = {\n  response: string;\n};\n\ninterface PerplexityResponse {\n  choices: Array<{\n    message: {\n      content: string;\n    };\n  }>;\n}\nexport type Run<C extends Record<string, any>, I extends Record<string, any>, R extends Record<string, any>> = (config: C, inputs: I) => Promise<R>;\n\nexport const run: Run<Configurations, Parameters, Result> = async (\n  configurations,\n  parameters,\n): Promise<Result> => {\n  const response = await fetch('https://api.perplexity.ai/chat/completions', {\n    method: 'POST',\n    headers: {\n      accept: 'application/json',\n      authorization: `Bearer ${configurations.apiKey}`,\n      'content-type': 'application/json',\n    },\n    body: JSON.stringify({\n      model: 'llama-3.1-sonar-small-128k-online',\n      messages: [\n        { role: 'system', content: 'Be precise and concise.' },\n        { role: 'user', content: parameters.query },\n      ],\n    }),\n  });\n\n  if (!response.ok) {\n    throw new Error(\n      `Failed to fetch data from Perplexity API, status: ${response.status}`,\n    );\n  }\n\n  const data = (await response.json()) as PerplexityResponse;\n  const responseContent =\n    data.choices[0]?.message?.content || 'No information available';\n\n  return { response: responseContent };\n};\n","tools":[],"config":[{"BasicConfig":{"key_name":"apiKey","description":"API key for accessing the Perplexity API","required":true,"type":null,"key_value":null}}],"description":"Searches the web using Perplexity API (limited)","keywords":["perplexity","api","shinkai"],"input_args":{"type":"object","properties":{"query":{"type":"string","description":"The search query to send to Perplexity API"}},"required":["query"]},"output_arg":{"json":""},"activated":false,"embedding":[0.11645089,0.3802397,-0.09272365,-0.67411584,-0.7537766,-0.032171614,-0.45825568,-0.49803934,-0.07670201,-0.057288826,-0.54977757,1.0430512,0.6679656,-0.1604064,0.4738902,0.50352037,0.21564056,-0.6655856,-1.5949762,0.22694373,0.5342862,0.7357455,0.50433016,0.28663254,-0.02867283,-0.031037897,-0.5205588,-0.041651927,-1.3853055,-2.3032627,0.8839041,0.30389038,-0.14362933,-0.2686656,-0.059750296,-0.5206286,-0.2615851,0.036210623,-0.6085326,0.055579424,0.29047012,0.47797203,-0.7003521,0.032415316,-0.29938808,-0.27016744,0.29302964,-0.19847909,0.53113383,0.3267147,-0.27189824,-0.04888197,0.18237285,0.34493575,-0.49145225,0.12325862,-0.092084035,-0.307699,-0.21803373,0.13233665,0.5443668,0.8112723,-3.989468,0.15532626,-0.15926404,0.3642894,-0.15024883,-0.41265762,0.10903634,0.7352435,0.11528505,-0.24735595,-0.21702892,0.22264877,0.064511694,-0.38259807,0.00055605546,0.14024082,0.14670368,-0.39417234,0.41306466,0.5847184,-0.3451556,-0.12406353,-0.032791402,0.43859747,-0.15003729,0.02542606,0.33477247,0.16663855,-0.703967,-0.1934891,0.03318881,-0.27462703,-0.733422,0.0606576,0.19128543,0.636479,0.5873402,3.5084991,0.45497346,-0.4058737,0.17792565,-0.29604372,0.6499731,-0.2301283,0.10245432,-0.1421148,-0.04591573,-0.06489993,-0.0492054,-0.06698479,0.45488554,0.41171885,-0.30030915,-0.59429204,0.1879906,0.30220774,0.20928462,0.5431731,-0.8490738,0.4382806,-0.37295023,-0.21139148,0.038392637,0.08154917,-1.0377465,0.27984643,0.43633434,-0.13672704,0.4028079,-0.26608434,-0.6147281,0.21044709,-0.124731585,0.09012708,0.35903013,-0.45551512,0.23613182,-0.6970096,0.3138734,-1.2784092,1.0144013,0.5601748,0.5553081,0.85539246,-0.3101067,-0.018381156,-0.36182493,0.2906356,-0.38740358,0.27574438,-0.106455676,-0.5253885,0.3617983,-0.27406377,0.10209672,0.05118013,-0.62402046,0.31799307,-0.22534356,0.17344433,0.50287616,0.8384853,0.4648925,-0.49320915,0.23515385,-0.04215353,0.12130138,-0.36237866,0.33668444,0.13787805,-0.029066041,0.44245827,-0.5305936,0.27351266,-0.771116,0.13033041,-0.08559129,-1.0153216,0.23239507,0.96706855,-0.021236736,-0.75734127,-0.15587541,0.28624126,0.2334207,-0.13497245,0.52321583,0.31058794,-0.5444632,2.087078,-0.7947092,-0.47221297,0.041387036,-0.19452813,-0.5043615,0.49735773,-0.17028435,-0.34137905,-0.3128182,-0.7499987,-0.018927649,0.07509248,0.14645281,-0.72122896,0.6193227,-0.55968297,0.42440146,-0.60465604,-0.0012502596,-0.20697278,0.017538324,0.0798682,0.5066283,-0.120684266,0.2885825,0.11684653,0.1198242,0.38593853,0.12598982,-0.18813714,-0.4122072,-0.492496,-0.4052348,-0.25425065,-0.477075,0.06350674,-0.28631818,-0.07108346,0.55995566,0.49807364,0.65559477,0.89627963,0.8168491,-0.20774716,-0.42542398,0.7693567,0.33275378,-0.45456913,0.5645004,-0.13026272,-0.2622282,-0.3358931,0.5276109,-1.1812775,0.015307136,0.06622091,-0.5011313,2.07899,1.0236517,-0.02029689,0.29381785,-0.032027494,-0.22639114,-0.13027939,-1.6197495,-0.047988273,-0.6458405,0.58243275,-0.38908395,-0.17550264,0.35894585,-0.07727822,0.19970895,0.0067867376,-0.86764807,-0.5372001,0.03185998,-0.25801203,-0.110302955,0.67893493,-0.098150775,0.44518924,0.2846071,-0.06250328,0.57811636,-0.017079383,-0.8492406,0.11247228,0.2989498,0.34395933,0.7322431,-0.013522863,-0.3790063,-0.38699606,-0.4408811,0.6623414,-0.55028033,0.08520629,-0.5261909,0.18402901,-0.5693831,-0.25149646,1.8593986,0.45961162,0.61170405,0.21321318,0.070005484,0.3738091,-0.23829171,0.064711,0.49576297,0.14187577,-0.13750578,-0.08783659,0.58597755,-0.34088972,-0.49333525,-0.19943738,-0.1679697,0.21279567,0.056609992,-0.6321387,0.46428663,-0.41508734,0.14146604,0.52955735,0.07605636,-1.9218364,-0.30150202,0.42497358,-0.006485235,-0.44263524,-0.28049743,0.42112368,-0.3480398,0.09213908,0.22300623,1.6682819,0.59545535,-0.45685238,-0.4643534,-0.48270938,0.72018564,-0.39325646,-0.14966555,-0.3658264,-0.7589981,-0.1770872,0.51835907,1.179172,0.6406662,0.6497778,-0.081445806,0.27046713,-0.32926446,-1.228502,0.09208644,0.33586162,-0.47407374,0.7246797,-0.029335575,-0.55128133,1.0291284,0.737099,0.37892345,-0.120127216,-0.25902197,1.7916461,0.08593016,-0.014039258,-0.3299547,0.38398552,0.101137094,0.56588775,-0.17234516,-0.2501452,-0.13663173,-0.09309722,-0.19804183,-0.6805411,0.29109603,0.1611939,0.041646678,0.4484455,0.07696198,0.41209462,0.37516358,0.34763062,-0.32356164,0.09036073,-0.6707306,0.41546535],"result":{"type":"object","properties":{"response":{"description":"The search results and analysis from Perplexity API","type":"string"}},"required":["response"]},"sql_tables":null,"sql_queries":null,"file_inbox":null,"oauth":null,"assets":null},false]}
