+++ tools/perplexity/index.test.ts
@@ -1,23 +1,20 @@
-import { expect } from 'jsr:@std/expect/expect';
-import { definition, run } from './index.ts';
+// import { expect } from 'jsr:@std/expect/expect';
+// import { run } from './tool.ts';
 
-Deno.test('exists definition', () => {
-  expect(definition).toBeInstanceOf(Object);
-});
 
-Deno.test({
-  name: 'run',
-  // Perplexity web scrapping is not working in CI (need to figure out why)
-  ignore: Deno.env.get('CI') === 'true',
-  fn: async () => {
-    const run_result = await run(
-      {
-        chromePath: Deno.env.get('CHROME_PATH'),
-      },
-      {
-        query: 'What is the meaning of life?',
-      },
-    );
-    expect(run_result.response).toEqual(expect.any(String));
-  },
-});
+// Deno.test({
+//   name: 'run',
+//   // Perplexity web scrapping is not working in CI (need to figure out why)
+//   ignore: Deno.env.get('CI') === 'true',
+//   fn: async () => {
+//     const run_result = await run(
+//       {
+//         chromePath: Deno.env.get('CHROME_PATH'),
+//       },
+//       {
+//         query: 'What is the meaning of life?',
+//       },
+//     );
+//     expect(run_result.response).toEqual(expect.any(String));
+//   },
+// });
