+++ tools/perplexity/metadata.json
@@ -3,6 +3,7 @@
     "name": "Perplexity",
     "description": "Searches the internet using Perplexity",
     "author": "Shinkai",
+    "version": "1.0.0",
     "keywords": [
       "perplexity",
       "shinkai"
