+++ tools/youtube-summary/.tool-dump.test.json
@@ -1,0 +1,1 @@
+{"type":"Deno","content":[{"toolkit_name":"youtube-summary","name":"Youtube Transcript Summarizer","author":"@@eduardosotomontaner.arb-sep-shinkai","version":"1.0.0","js_code":"import { YoutubeTranscript } from 'npm:youtube-transcript@1.2.1';\nimport { shinkaiLlmPromptProcessor } from './shinkai-local-tools.ts'\n\n// Tool does not need any configuration\ntype CONFIG = Record<string, unknown>;\n\ntype INPUTS = {\n  url: string;\n  lang?: string;\n};\n\ntype OUTPUT = { summary: string };\n\nexport const run: Run<CONFIG, INPUTS, OUTPUT> = async (\n  _configurations,\n  parameters,\n): Promise<OUTPUT> => {\n  console.log(`transcripting ${parameters.url}`);\n\n  // Get transcription\n  const transcript = await YoutubeTranscript.fetchTranscript(parameters.url, {\n    lang: parameters.lang || 'en',\n  });\n\n  // Send to LLM to build a formatted response\n  const message = `\n    According to this transcription of a youtube video (which is in csv separated by ':::'):\n\n    offset;text\n    ${transcript.map((v) => `${Math.floor(v.offset)}:::${v.text}`).join('\\n')}\n    ---------------\n\n    The video URL is ${parameters.url}\n\n    ---------------\n\n    Write a detailed summary divided in sections along the video.\n    Format the answer using markdown.\n    Add markdown links referencing every section using this format https://www.youtube.com/watch?v={video_id}&t={offset} where 'offset' is a number and can be obtained from the transcription in csv format to generate the URL\n  `;\n  const response = await shinkaiLlmPromptProcessor({ format: 'text', prompt: message })\n  return { summary: response.message }\n};\n","tools":["local:::rust_toolkit:::shinkai_llm_prompt_processor"],"config":[],"description":"Fetches the transcript of a YouTube video and generates a formatted summary using an LLM.","keywords":["youtube","transcript","summary","video","LLM"],"input_args":{"type":"object","properties":{"lang":{"type":"string","description":"The language for the transcript (optional)"},"url":{"type":"string","description":"The URL of the YouTube video"}},"required":["url"]},"output_arg":{"json":""},"activated":false,"embedding":[0.24397473,0.19737607,0.10249484,0.07853881,-0.52260065,-0.07264366,-1.4292128,0.5816098,-0.0569295,0.467394,-0.2629992,1.546936,0.030428365,0.18555912,0.5241776,0.40233222,-0.10415818,-0.7092806,-1.297857,-0.664604,0.52484876,0.5144266,-0.0060532093,-0.0029968247,0.13401021,-0.13712981,0.09876527,-0.22174498,-1.0573201,-1.9738672,0.70428807,0.54259884,-0.16001034,-0.37234327,-0.43517393,-1.3643596,0.25747937,0.46092734,-0.292954,0.4579779,-0.09054327,0.13232066,-0.7888382,0.40833142,-0.043910377,-0.19611594,0.36788312,-0.09296358,1.4296865,0.44269896,-0.40729654,-0.19846602,-0.10134886,0.49915153,-0.17917651,-0.61800796,-0.39331928,-0.5531162,0.93573284,0.3125194,-0.032507822,-0.024462327,-3.7117436,-0.23966308,-0.082733504,-0.00083688647,-0.2638889,-0.079310715,-0.04274538,0.38685042,-1.0061314,-0.09795482,-0.595692,0.8140711,-0.52223,-0.5010422,0.48327318,-0.30269998,0.26365638,-0.53166157,0.06926101,0.061586447,0.22167054,-0.38880783,-0.53398055,0.90134996,-0.47085276,-0.24938089,-0.13700266,0.26321328,-0.9187829,-0.030944848,0.6005912,0.34225106,-0.54002047,-0.3963846,0.6813786,-0.4724697,0.072636716,2.6400874,1.0521492,0.014715459,0.1956683,-1.1673645,0.4569034,-0.70678335,0.47124594,-0.42230985,0.31645888,0.0359073,0.78643477,-0.17552736,0.15710464,0.549718,0.51373,0.37795103,-0.81747603,0.024700543,0.03176478,0.36899906,-0.272966,-0.12973304,-0.39982963,-0.18365073,-0.46064875,0.14623581,-0.1908139,0.15239067,0.5717435,-0.7718099,-0.041600056,-0.055492602,-0.6733642,0.6702037,0.025288485,0.30686027,0.40308362,-0.8706652,0.31286785,-0.81772375,-0.18790466,-2.0614433,0.6394961,-0.018431015,0.34960693,-0.007028464,-0.61032754,0.37297478,-0.29247496,-0.465864,0.36508018,0.6427132,0.22788927,-0.08449084,0.88174784,-0.13330701,-0.72458035,-0.55865467,-0.20877361,0.050680257,-0.38116446,-0.07186862,0.624656,-0.3708253,0.2717578,-0.5983692,0.77082396,0.16514742,0.39356315,0.32194296,0.28758615,0.11886267,0.01459287,0.7468215,-0.025611265,-0.89384717,-0.4739496,-0.301784,-0.0015290305,0.29546043,0.48127237,0.2063696,-0.409744,-0.04899437,-0.23912765,0.6469834,0.09129079,0.24771339,0.96164113,0.7420727,-0.33983496,1.6778651,-1.1015447,-0.17468059,-0.111047395,0.09020809,-0.3394488,0.3813395,0.30324456,0.06054409,-0.9422036,-0.21234436,0.25479048,0.0029522553,-0.33134744,-0.64197415,-0.31880897,0.73515236,-0.370654,-0.61103266,0.49827653,-0.9070959,0.5150069,0.54918575,-0.0791596,-0.26136273,-0.038667,0.20681158,-0.7998123,0.15697767,0.31832293,-0.30024567,-0.13398749,-0.5464064,0.078004286,0.5013751,-0.5964775,0.7934169,0.18798716,-0.102526814,0.7516279,1.3288732,0.38466543,1.1970915,0.70729077,0.16436699,-0.4109063,0.31046456,-0.09760972,-0.2675225,0.18276238,0.25188726,0.069069535,0.28543174,0.2071175,-0.08041818,-0.6181666,0.043410942,-0.29807296,1.7043421,0.43463537,0.4075363,0.59444076,0.82585335,0.6737303,-0.36265337,-1.4637611,-0.29978368,-0.20907988,0.33043876,-0.750395,0.6137083,0.19155118,0.2934378,-0.5209952,-0.36604527,-0.69852865,-0.4371357,0.028415933,0.18738863,-0.30432528,0.49997562,-0.11741397,0.042816848,-0.4490757,0.4281727,0.18414852,0.42120722,-0.44500935,-0.9356021,0.004114732,-0.098933734,0.20038745,0.31341612,-0.7175617,-0.13790105,-0.078082725,-0.30331305,0.016176544,0.5784527,0.09730462,-0.46089923,-0.9954212,0.2812389,1.5212535,0.25866175,0.15545672,0.5862159,0.3469371,0.09966384,-0.93005514,0.027373465,-0.14089948,-0.49164492,-0.18993902,-0.6721664,0.91621786,-0.31503543,-0.43121752,0.4722464,-0.42334154,0.41654542,0.2577617,-0.2962021,0.46720228,-0.3402864,0.3280484,0.9959594,-0.17491736,-1.8942695,-0.035868153,-0.29466665,-0.25931674,-0.27141005,0.24185471,1.0774001,-0.64388967,0.6423959,0.24093969,1.5087445,0.11422682,-0.10210596,-0.35378647,0.08158391,0.7351494,-0.47266462,-0.18411069,-0.30488995,-0.7604541,-0.2951124,0.32563996,1.0398327,0.033854738,1.0128971,0.04701845,-0.2543579,-0.3876006,-1.0073549,0.6400236,-0.24982655,-0.3674008,0.75311905,-0.40171337,-0.31875148,0.7237389,0.8828592,0.5095583,-0.050843228,0.3896459,1.7165403,0.49974895,-0.35941413,0.277945,0.12831461,0.41759193,0.11227698,0.021763999,-0.5772322,0.04317443,-0.34799072,0.57283884,-0.45722812,0.7255814,0.7775589,0.109948486,0.2851043,0.6291014,0.62423563,0.36268318,-0.022946805,-0.10817164,-0.037535723,-0.21269107,-0.74232906],"result":{"type":"object","properties":{"summary":{"description":"The generated summary of the video","type":"string"}},"required":["summary"]},"sql_tables":[],"sql_queries":[],"file_inbox":null,"oauth":null,"assets":null},false]}
