+++ tools/youtube-summary/index.test.ts
@@ -1,20 +1,20 @@
-import { expect } from 'jsr:@std/expect/expect';
-import { run } from './tool.ts';
+// import { expect } from 'jsr:@std/expect/expect';
+// import { run } from './tool.ts';
 
-Deno.test({
-  name: 'transcript video',
-  ignore: Deno.env.get('CI') === 'true',
-  fn: async () => {
-    const result = await run(
-      {
-        // No configuration needed
-      },
-      {
-        url: 'https://www.youtube.com/watch?v=SUj34OWkjXU',
-        lang: 'en',
-      },
-    );
-    expect(result.summary.length).toBeGreaterThan(0);
-    console.log(result.summary);
-  },
-});
+// Deno.test({
+//   name: 'transcript video',
+//   ignore: Deno.env.get('CI') === 'true',
+//   fn: async () => {
+//     const result = await run(
+//       {
+//         // No configuration needed
+//       },
+//       {
+//         url: 'https://www.youtube.com/watch?v=SUj34OWkjXU',
+//         lang: 'en',
+//       },
+//     );
+//     expect(result.summary.length).toBeGreaterThan(0);
+//     console.log(result.summary);
+//   },
+// });
